@@ -11,12 +11,10 @@
     "format": "prettier --write ."
   },
   "devDependencies": {
-    "@skeletonlabs/skeleton": "2.10.2",
-    "@skeletonlabs/tw-plugin": "0.4.0",
-    "@sveltejs/adapter-auto": "^3.0.0",
+    "@sveltejs/adapter-auto": "^3.3.1",
     "@sveltejs/adapter-node": "^5.2.8",
-    "@sveltejs/kit": "^2.0.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/kit": "^2.5.27",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@types/node": "22.7.4",
     "autoprefixer": "^10.4.20",
     "bits-ui": "^0.21.16",
@@ -26,16 +24,16 @@
     "mode-watcher": "^0.4.1",
     "paneforge": "^0.0.6",
     "postcss": "8.4.47",
-    "svelte": "^4.2.7",
-    "svelte-check": "^3.6.0",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
     "svelte-sonner": "^0.3.28",
     "sveltekit-superforms": "^2.20.0",
     "tailwind-merge": "^2.5.4",
     "tailwind-variants": "^0.2.1",
     "tailwindcss": "^3.4.9",
     "tslib": "^2.4.1",
-    "typescript": "^5.0.0",
-    "vite": "^5.0.3",
+    "typescript": "^5.5.0",
+    "vite": "^5.4.4",
     "vite-plugin-tailwind-purgecss": "0.3.3",
     "zod": "^3.23.8"
   },
