@@ -4,15 +4,20 @@
 
     type $$Props = AvatarPrimitive.Props;
 
-    let className: $$Props["class"] = undefined;
-    export let delayMs: $$Props["delayMs"] = undefined;
-    export { className as class };
+    interface Props {
+        class?: $$Props["class"];
+        delayMs?: $$Props["delayMs"];
+        children?: import("svelte").Snippet;
+        [key: string]: any;
+    }
+
+    let { class: className = undefined, delayMs = undefined, children, ...rest }: Props = $props();
 </script>
 
 <AvatarPrimitive.Root
     {delayMs}
     class={cn("relative flex h-10 w-10 shrink-0 overflow-hidden rounded-full", className)}
-    {...$$restProps}
+    {...rest}
 >
-    <slot />
+    {@render children?.()}
 </AvatarPrimitive.Root>
