@@ -8,13 +8,25 @@
         hideCloseButton?: boolean;
     };
 
-    let className: $$Props["class"] = undefined;
-    export let transition: $$Props["transition"] = flyAndScale;
-    export let transitionConfig: $$Props["transitionConfig"] = {
-        duration: 200
-    };
-    export let hideCloseButton: $$Props["hideCloseButton"] = false;
-    export { className as class };
+    interface Props {
+        class?: $$Props["class"];
+        transition?: $$Props["transition"];
+        transitionConfig?: $$Props["transitionConfig"];
+        hideCloseButton?: $$Props["hideCloseButton"];
+        children?: import("svelte").Snippet;
+        [key: string]: any;
+    }
+
+    let {
+        class: className = undefined,
+        transition = flyAndScale,
+        transitionConfig = {
+            duration: 200
+        },
+        hideCloseButton = false,
+        children,
+        ...rest
+    }: Props = $props();
 </script>
 
 <Dialog.Portal>
@@ -26,9 +38,9 @@
             "fixed left-[50%] top-[50%] z-50 grid w-full max-w-lg translate-x-[-50%] translate-y-[-50%] gap-4 border bg-background p-6 shadow-lg sm:rounded-lg md:w-full",
             className
         )}
-        {...$$restProps}
+        {...rest}
     >
-        <slot />
+        {@render children?.()}
         {#if !hideCloseButton}
             <DialogPrimitive.Close
                 class="absolute right-4 top-4 rounded-sm opacity-70 ring-offset-background transition-opacity hover:opacity-100 focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:pointer-events-none data-[state=open]:bg-accent data-[state=open]:text-muted-foreground"
