@@ -4,13 +4,18 @@
 
     type $$Props = HTMLAttributes<HTMLDivElement>;
 
-    let className: $$Props["class"] = undefined;
-    export { className as class };
+    interface Props {
+        class?: $$Props["class"];
+        children?: import("svelte").Snippet;
+        [key: string]: any;
+    }
+
+    let { class: className = undefined, children, ...rest }: Props = $props();
 </script>
 
 <div
     class={cn("flex flex-col-reverse sm:flex-row sm:justify-end sm:space-x-2", className)}
-    {...$$restProps}
+    {...rest}
 >
-    <slot />
+    {@render children?.()}
 </div>
