@@ -4,10 +4,15 @@
 
     type $$Props = HTMLAttributes<HTMLDivElement>;
 
-    let className: $$Props["class"] = undefined;
-    export { className as class };
+    interface Props {
+        class?: $$Props["class"];
+        children?: import("svelte").Snippet;
+        [key: string]: any;
+    }
+
+    let { class: className = undefined, children, ...rest }: Props = $props();
 </script>
 
-<div class={cn("flex flex-col space-y-1.5 text-center sm:text-left", className)} {...$$restProps}>
-    <slot />
+<div class={cn("flex flex-col space-y-1.5 text-center sm:text-left", className)} {...rest}>
+    {@render children?.()}
 </div>
