@@ -1,4 +1,7 @@
 <script lang="ts">
+    import { createBubbler, passive } from "svelte/legacy";
+
+    const bubble = createBubbler();
     import type { HTMLInputAttributes } from "svelte/elements";
     import type { InputEvents } from "./index.js";
     import { cn } from "$lib/utils.js";
@@ -6,13 +9,21 @@
     type $$Props = HTMLInputAttributes;
     type $$Events = InputEvents;
 
-    let className: $$Props["class"] = undefined;
-    export let value: $$Props["value"] = undefined;
-    export { className as class };
-
     // Workaround for https://github.com/sveltejs/svelte/issues/9305
     // Fixed in Svelte 5, but not backported to 4.x.
-    export let readonly: $$Props["readonly"] = undefined;
+    interface Props {
+        class?: $$Props["class"];
+        value?: $$Props["value"];
+        readonly?: $$Props["readonly"];
+        [key: string]: any;
+    }
+
+    let {
+        class: className = undefined,
+        value = $bindable(undefined),
+        readonly = undefined,
+        ...rest
+    }: Props = $props();
 </script>
 
 <input
@@ -22,21 +33,21 @@
     )}
     bind:value
     {readonly}
-    on:blur
-    on:change
-    on:click
-    on:focus
-    on:focusin
-    on:focusout
-    on:keydown
-    on:keypress
-    on:keyup
-    on:mouseover
-    on:mouseenter
-    on:mouseleave
-    on:mousemove
-    on:paste
-    on:input
-    on:wheel|passive
-    {...$$restProps}
+    onblur={bubble("blur")}
+    onchange={bubble("change")}
+    onclick={bubble("click")}
+    onfocus={bubble("focus")}
+    onfocusin={bubble("focusin")}
+    onfocusout={bubble("focusout")}
+    onkeydown={bubble("keydown")}
+    onkeypress={bubble("keypress")}
+    onkeyup={bubble("keyup")}
+    onmouseover={bubble("mouseover")}
+    onmouseenter={bubble("mouseenter")}
+    onmouseleave={bubble("mouseleave")}
+    onmousemove={bubble("mousemove")}
+    onpaste={bubble("paste")}
+    oninput={bubble("input")}
+    use:passive={["wheel", () => bubble("wheel")]}
+    {...rest}
 />
