@@ -5,8 +5,13 @@
     type $$Props = LabelPrimitive.Props;
     type $$Events = LabelPrimitive.Events;
 
-    let className: $$Props["class"] = undefined;
-    export { className as class };
+    interface Props {
+        class?: $$Props["class"];
+        children?: import("svelte").Snippet;
+        [key: string]: any;
+    }
+
+    let { class: className = undefined, children, ...rest }: Props = $props();
 </script>
 
 <LabelPrimitive.Root
@@ -14,8 +19,8 @@
         "text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70",
         className
     )}
-    {...$$restProps}
+    {...rest}
     on:mousedown
 >
-    <slot />
+    {@render children?.()}
 </LabelPrimitive.Root>
