@@ -5,9 +5,14 @@
     type $$Props = TabsPrimitive.TriggerProps;
     type $$Events = TabsPrimitive.TriggerEvents;
 
-    let className: $$Props["class"] = undefined;
-    export let value: $$Props["value"];
-    export { className as class };
+    interface Props {
+        class?: $$Props["class"];
+        value: $$Props["value"];
+        children?: import("svelte").Snippet;
+        [key: string]: any;
+    }
+
+    let { class: className = undefined, value, children, ...rest }: Props = $props();
 </script>
 
 <TabsPrimitive.Trigger
@@ -16,8 +21,8 @@
         className
     )}
     {value}
-    {...$$restProps}
+    {...rest}
     on:click
 >
-    <slot />
+    {@render children?.()}
 </TabsPrimitive.Trigger>
