@@ -1,17 +1,21 @@
-<script>
-    import { Button } from "$components/ui/button";
+<script lang="ts">
+    interface Props {
+        title?: string;
+        description?: string;
+    }
+
+    let { title, description }: Props = $props();
+    const titleFallback = "Problem Title";
+    const descriptionFallback = `
+        <p>
+            Code Minecraft in O(1) bro (Just kidding. Please wait for problem to load :D)
+        </p>
+    `;
 </script>
 
 <div class="h-full bg-background px-4 py-5">
-    <h2 class="mb-4 text-2xl font-bold">Binary Search</h2>
-    <p class="mb-4">
-        Implement a binary search algorithm to find the position of a target value within a sorted
-        array.
-    </p>
-    <h3 class="mb-2 text-xl font-semibold">Example:</h3>
-    <pre class="rounded bg-secondary p-2">
-        Input: nums = [-1,0,3,5,9,12], target = 9
-        Output: 4
-        Explanation: 9 exists in nums and its index is 4
-    </pre>
+    <h2 class="mb-4 text-2xl font-bold">{title}</h2>
+    <div>
+        {@html description}
+    </div>
 </div>
