@@ -5,17 +5,26 @@
 
     type $$Props = DialogPrimitive.OverlayProps;
 
-    let className: $$Props["class"] = undefined;
-    export let transition: $$Props["transition"] = fade;
-    export let transitionConfig: $$Props["transitionConfig"] = {
-        duration: 150
-    };
-    export { className as class };
+    interface Props {
+        class?: $$Props["class"];
+        transition?: $$Props["transition"];
+        transitionConfig?: $$Props["transitionConfig"];
+        [key: string]: any;
+    }
+
+    let {
+        class: className = undefined,
+        transition = fade,
+        transitionConfig = {
+            duration: 150
+        },
+        ...rest
+    }: Props = $props();
 </script>
 
 <DialogPrimitive.Overlay
     {transition}
     {transitionConfig}
     class={cn("fixed inset-0 z-50 bg-background/80 backdrop-blur-sm", className)}
-    {...$$restProps}
+    {...rest}
 />
