@@ -1,10 +1,10 @@
 <script lang="ts">
-    import * as Button from "$lib/components/ui/button/index.js";
+	import * as Button from "$lib/components/ui/button/index.js";
 
-    type $$Props = Button.Props;
-    type $$Events = Button.Events;
+	type $$Props = Button.Props;
+	type $$Events = Button.Events;
 </script>
 
 <Button.Root type="submit" on:click on:keydown {...$$restProps}>
-    <slot />
+	<slot />
 </Button.Root>
