@@ -1,17 +1,17 @@
 <script lang="ts">
-    import * as FormPrimitive from "formsnap";
-    import type { HTMLAttributes } from "svelte/elements";
-    import { cn } from "$lib/utils.js";
+	import * as FormPrimitive from "formsnap";
+	import type { HTMLAttributes } from "svelte/elements";
+	import { cn } from "$lib/utils.js";
 
-    type $$Props = HTMLAttributes<HTMLSpanElement>;
-    let className: string | undefined | null = undefined;
-    export { className as class };
+	type $$Props = HTMLAttributes<HTMLSpanElement>;
+	let className: string | undefined | null = undefined;
+	export { className as class };
 </script>
 
 <FormPrimitive.Description
-    class={cn("text-sm text-muted-foreground", className)}
-    {...$$restProps}
-    let:descriptionAttrs
+	class={cn("text-muted-foreground text-sm", className)}
+	{...$$restProps}
+	let:descriptionAttrs
 >
-    <slot {descriptionAttrs} />
+	<slot {descriptionAttrs} />
 </FormPrimitive.Description>
