@@ -1,25 +1,25 @@
 <script lang="ts" context="module">
-    import type { FormPathLeaves, SuperForm } from "sveltekit-superforms";
-    type T = Record<string, unknown>;
-    type U = FormPathLeaves<T>;
+	import type { FormPathLeaves, SuperForm } from "sveltekit-superforms";
+	type T = Record<string, unknown>;
+	type U = FormPathLeaves<T>;
 </script>
 
 <script lang="ts" generics="T extends Record<string, unknown>, U extends FormPathLeaves<T>">
-    import type { HTMLAttributes } from "svelte/elements";
-    import * as FormPrimitive from "formsnap";
-    import { cn } from "$lib/utils.js";
+	import type { HTMLAttributes } from "svelte/elements";
+	import * as FormPrimitive from "formsnap";
+	import { cn } from "$lib/utils.js";
 
-    type $$Props = FormPrimitive.ElementFieldProps<T, U> & HTMLAttributes<HTMLElement>;
+	type $$Props = FormPrimitive.ElementFieldProps<T, U> & HTMLAttributes<HTMLElement>;
 
-    export let form: SuperForm<T>;
-    export let name: U;
+	export let form: SuperForm<T>;
+	export let name: U;
 
-    let className: $$Props["class"] = undefined;
-    export { className as class };
+	let className: $$Props["class"] = undefined;
+	export { className as class };
 </script>
 
 <FormPrimitive.ElementField {form} {name} let:constraints let:errors let:tainted let:value>
-    <div class={cn("space-y-2", className)}>
-        <slot {constraints} {errors} {tainted} {value} />
-    </div>
+	<div class={cn("space-y-2", className)}>
+		<slot {constraints} {errors} {tainted} {value} />
+	</div>
 </FormPrimitive.ElementField>
