@@ -1,30 +1,30 @@
 <script lang="ts" context="module">
-    import type { FormPath, SuperForm } from "sveltekit-superforms";
-    type T = Record<string, unknown>;
-    type U = FormPath<T>;
+	import type { FormPath, SuperForm } from "sveltekit-superforms";
+	type T = Record<string, unknown>;
+	type U = FormPath<T>;
 </script>
 
 <script lang="ts" generics="T extends Record<string, unknown>, U extends FormPath<T>">
-    import * as FormPrimitive from "formsnap";
-    import { cn } from "$lib/utils.js";
+	import * as FormPrimitive from "formsnap";
+	import { cn } from "$lib/utils.js";
 
-    type $$Props = FormPrimitive.FieldsetProps<T, U>;
+	type $$Props = FormPrimitive.FieldsetProps<T, U>;
 
-    export let form: SuperForm<T>;
-    export let name: U;
+	export let form: SuperForm<T>;
+	export let name: U;
 
-    let className: $$Props["class"] = undefined;
-    export { className as class };
+	let className: $$Props["class"] = undefined;
+	export { className as class };
 </script>
 
 <FormPrimitive.Fieldset
-    {form}
-    {name}
-    let:constraints
-    let:errors
-    let:tainted
-    let:value
-    class={cn("space-y-2", className)}
+	{form}
+	{name}
+	let:constraints
+	let:errors
+	let:tainted
+	let:value
+	class={cn("space-y-2", className)}
 >
-    <slot {constraints} {errors} {tainted} {value} />
+	<slot {constraints} {errors} {tainted} {value} />
 </FormPrimitive.Fieldset>
