@@ -1,17 +1,17 @@
 <script lang="ts">
-    import * as FormPrimitive from "formsnap";
-    import { cn } from "$lib/utils.js";
+	import * as FormPrimitive from "formsnap";
+	import { cn } from "$lib/utils.js";
 
-    type $$Props = FormPrimitive.LegendProps;
+	type $$Props = FormPrimitive.LegendProps;
 
-    let className: $$Props["class"] = undefined;
-    export { className as class };
+	let className: $$Props["class"] = undefined;
+	export { className as class };
 </script>
 
 <FormPrimitive.Legend
-    {...$$restProps}
-    class={cn("text-sm font-medium leading-none data-[fs-error]:text-destructive", className)}
-    let:legendAttrs
+	{...$$restProps}
+	class={cn("data-[fs-error]:text-destructive text-sm font-medium leading-none", className)}
+	let:legendAttrs
 >
-    <slot {legendAttrs} />
+	<slot {legendAttrs} />
 </FormPrimitive.Legend>
